# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Px Lite Production API
  version: 1.0.0
  description: >-
    Px Lite 医药患教内容运营与聚合行为洞察 API。当前契约基于代码现状整理；财务模块仍为前端 demo-only，暂未提供生产 API。
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /distribution/projects:
    get:
      summary: 获取分发项目列表
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 分发项目分页列表
  /distribution/projects/{id}:
    get:
      summary: 获取分发项目详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 分发项目详情
        '404':
          description: 分发项目不存在
  /distribution/projects/{id}/doctors:
    get:
      summary: 获取分发项目医生候选
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 医生候选列表
  /distribution/requests/{id}:
    get:
      summary: 获取诉求维度分发工作台
      description: 返回选题诉求、诉求级分发配置、医生候选与历史分发批次。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 诉求维度分发工作台数据
        '404':
          description: 诉求不存在
  /distribution/requests/{id}/accept:
    post:
      summary: 受理诉求并进入拆单/派单流程
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
      responses:
        '200':
          description: 诉求已受理
        '403':
          description: 无分发写权限
        '404':
          description: 诉求不存在
  /distribution/requests/{id}/config:
    put:
      summary: 保存诉求级医生/患者分发策略
      description: 写入 `request_distribution_configs`，包括医生筛选、指定医生额度、患者触达渠道、灰度比例和患者上限。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 策略已保存
        '403':
          description: 无分发写权限
        '404':
          description: 诉求不存在
  /distribution/requests/{id}/batches:
    post:
      summary: 提交诉求分发批次
      description: 写入 `request_distribution_batches`，记录本次主题 × 形式篇数、指定医生篇数和策略自动篇数。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 分发批次已创建
        '400':
          description: 诉求不存在或批次为空
        '403':
          description: 无分发写权限
  /distribution:
    get:
      summary: 获取分发策略列表
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 分发策略分页列表
    post:
      summary: 创建分发策略
      responses:
        '201':
          description: 分发策略已创建
        '403':
          description: 无分发写权限
  /distribution/{id}:
    put:
      summary: 更新分发策略
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 分发策略已更新
        '404':
          description: 分发策略不存在
...